Add missing cast to fix divide by zero exception in DefaultProgress::Report
- time_from_epoch was not casting tv.tv_sec to a uint64_t, which led to
returning values like 0xffffffff800bb69c. This would lead
DefaultProgress::Report to encounter a divide by zero error
approximately every 4294 seconds.

+++ aff4/aff4_utils.h
@@ -34,7 +34,7 @@ std::string aff4_sprintf(std::string fmt, ...);
 inline uint64_t time_from_epoch() {
     struct timeval tv;
     gettimeofday(&tv, nullptr);
-    return tv.tv_sec * 1000000 + tv.tv_usec;
+    return ((uint64_t) tv.tv_sec * 1000000) + tv.tv_usec;
 }
 
 std::string GetLastErrorMessage();
